+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * `IconButton` updates:
   - The `color` property is now `backgroundColor`
   - The widget now takes a `Widget icon` rather than `IconData iconData` for better control over widget properties
+  - Deprecate and remove internal `foregroundColor` value
 
 ## [0.2.0]
 * New widget: `BackButton`, `IconButton`